Rofi integration partially working
news can be selected in rofi, then its body is printed in rofi.
The body needs to be formated, not done yet
Code quality needs to be better

it would be a good idea to implement the shell call to open the news
in the terminal, instead of rofi, to let the user decide.